ファイル分割した（これ以降はfreeをどうするか決めてから)
